[IMP] Add the possibility to have more than 1 vehicle by PO
Initially, there was only one vehicle for one PO.
Now, if the vehicle is set on the PO, all the PO lines are related to this vehicle
But if the vehicle is not set on the PO, the user can decide to have different vehicles on every line.
Use case is a grouped purchase of several items for several vehicles.